# Refactor to ES6 classes