--- a/lib/contentTypeParser.js
+++ b/lib/contentTypeParser.js
@@ -29,126 +29,128 @@ const {
 } = require('./errors')
 const warning = require('./warnings')
 
-function ContentTypeParser (bodyLimit, onProtoPoisoning, onConstructorPoisoning) {
-  this[kDefaultJsonParse] = getDefaultJsonParser(onProtoPoisoning, onConstructorPoisoning)
-  this.customParsers = {}
-  this.customParsers['application/json'] = new Parser(true, false, bodyLimit, this[kDefaultJsonParse])
-  this.customParsers['text/plain'] = new Parser(true, false, bodyLimit, defaultPlainTextParser)
-  this.parserList = ['application/json', 'text/plain']
-  this.parserRegExpList = []
-  this.cache = lru(100)
-}
-
-ContentTypeParser.prototype.add = function (contentType, opts, parserFn) {
-  const contentTypeIsString = typeof contentType === 'string'
-
-  if (!contentTypeIsString && !(contentType instanceof RegExp)) throw new FST_ERR_CTP_INVALID_TYPE()
-  if (contentTypeIsString && contentType.length === 0) throw new FST_ERR_CTP_EMPTY_TYPE()
-  if (typeof parserFn !== 'function') throw new FST_ERR_CTP_INVALID_HANDLER()
-
-  if (this.existingParser(contentType)) {
-    throw new FST_ERR_CTP_ALREADY_PRESENT(contentType)
+class ContentTypeParser {
+  constructor (bodyLimit, onProtoPoisoning, onConstructorPoisoning) {
+    this[kDefaultJsonParse] = getDefaultJsonParser(onProtoPoisoning, onConstructorPoisoning)
+    this.customParsers = {}
+    this.customParsers['application/json'] = new Parser(true, false, bodyLimit, this[kDefaultJsonParse])
+    this.customParsers['text/plain'] = new Parser(true, false, bodyLimit, defaultPlainTextParser)
+    this.parserList = ['application/json', 'text/plain']
+    this.parserRegExpList = []
+    this.cache = lru(100)
   }
 
-  if (opts.parseAs !== undefined) {
-    if (opts.parseAs !== 'string' && opts.parseAs !== 'buffer') {
-      throw new FST_ERR_CTP_INVALID_PARSE_TYPE(opts.parseAs)
+  add (contentType, opts, parserFn) {
+    const contentTypeIsString = typeof contentType === 'string'
+
+    if (!contentTypeIsString && !(contentType instanceof RegExp)) throw new FST_ERR_CTP_INVALID_TYPE()
+    if (contentTypeIsString && contentType.length === 0) throw new FST_ERR_CTP_EMPTY_TYPE()
+    if (typeof parserFn !== 'function') throw new FST_ERR_CTP_INVALID_HANDLER()
+
+    if (this.existingParser(contentType)) {
+      throw new FST_ERR_CTP_ALREADY_PRESENT(contentType)
     }
-  }
 
-  const parser = new Parser(
-    opts.parseAs === 'string',
-    opts.parseAs === 'buffer',
-    opts.bodyLimit,
-    parserFn
-  )
-
-  if (contentTypeIsString && contentType === '*') {
-    this.customParsers[''] = parser
-  } else {
-    if (contentTypeIsString) {
-      if (contentType !== 'application/json' && contentType !== 'text/plain') {
-        this.parserList.unshift(contentType)
+    if (opts.parseAs !== undefined) {
+      if (opts.parseAs !== 'string' && opts.parseAs !== 'buffer') {
+        throw new FST_ERR_CTP_INVALID_PARSE_TYPE(opts.parseAs)
       }
-    } else {
-      this.parserRegExpList.unshift(contentType)
     }
-    this.customParsers[contentType] = parser
-  }
-}
 
-ContentTypeParser.prototype.hasParser = function (contentType) {
-  return contentType in this.customParsers
-}
-
-ContentTypeParser.prototype.existingParser = function (contentType) {
-  if (contentType === 'application/json') {
-    return this.customParsers['application/json'].fn !== this[kDefaultJsonParse]
-  }
-  if (contentType === 'text/plain') {
-    return this.customParsers['text/plain'].fn !== defaultPlainTextParser
-  }
-
-  return contentType in this.customParsers
-}
-
-ContentTypeParser.prototype.getParser = function (contentType) {
-  // eslint-disable-next-line no-var
-  for (var i = 0; i !== this.parserList.length; ++i) {
-    const parserName = this.parserList[i]
-    if (contentType.indexOf(parserName) > -1) {
-      const parser = this.customParsers[parserName]
-      this.cache.set(contentType, parser)
-      return parser
-    }
-  }
-
-  // eslint-disable-next-line no-var
-  for (var j = 0; j !== this.parserRegExpList.length; ++j) {
-    const parserRegExp = this.parserRegExpList[j]
-    if (parserRegExp.test(contentType)) {
-      const parser = this.customParsers[parserRegExp]
-      this.cache.set(contentType, parser)
-      return parser
-    }
-  }
-
-  return this.customParsers['']
-}
-
-ContentTypeParser.prototype.run = function (contentType, handler, request, reply) {
-  const parser = this.cache.get(contentType) || this.getParser(contentType)
-
-  if (parser === undefined) {
-    reply.send(new FST_ERR_CTP_INVALID_MEDIA_TYPE(contentType))
-  } else if (parser.asString === true || parser.asBuffer === true) {
-    rawBody(
-      request,
-      reply,
-      reply.context._parserOptions,
-      parser,
-      done
+    const parser = new Parser(
+      opts.parseAs === 'string',
+      opts.parseAs === 'buffer',
+      opts.bodyLimit,
+      parserFn
     )
-  } else {
-    let result
 
-    if (parser.isDeprecatedSignature) {
-      result = parser.fn(request[kRequestPayloadStream], done)
+    if (contentTypeIsString && contentType === '*') {
+      this.customParsers[''] = parser
     } else {
-      result = parser.fn(request, request[kRequestPayloadStream], done)
-    }
-
-    if (result && typeof result.then === 'function') {
-      result.then(body => done(null, body), done)
+      if (contentTypeIsString) {
+        if (contentType !== 'application/json' && contentType !== 'text/plain') {
+          this.parserList.unshift(contentType)
+        }
+      } else {
+        this.parserRegExpList.unshift(contentType)
+      }
+      this.customParsers[contentType] = parser
     }
   }
 
-  function done (error, body) {
-    if (error) {
-      reply.send(error)
+  hasParser (contentType) {
+    return contentType in this.customParsers
+  }
+
+  existingParser (contentType) {
+    if (contentType === 'application/json') {
+      return this.customParsers['application/json'].fn !== this[kDefaultJsonParse]
+    }
+    if (contentType === 'text/plain') {
+      return this.customParsers['text/plain'].fn !== defaultPlainTextParser
+    }
+
+    return contentType in this.customParsers
+  }
+
+  getParser (contentType) {
+    // eslint-disable-next-line no-var
+    for (var i = 0; i !== this.parserList.length; ++i) {
+      const parserName = this.parserList[i]
+      if (contentType.indexOf(parserName) > -1) {
+        const parser = this.customParsers[parserName]
+        this.cache.set(contentType, parser)
+        return parser
+      }
+    }
+
+    // eslint-disable-next-line no-var
+    for (var j = 0; j !== this.parserRegExpList.length; ++j) {
+      const parserRegExp = this.parserRegExpList[j]
+      if (parserRegExp.test(contentType)) {
+        const parser = this.customParsers[parserRegExp]
+        this.cache.set(contentType, parser)
+        return parser
+      }
+    }
+
+    return this.customParsers['']
+  }
+
+  run (contentType, handler, request, reply) {
+    const parser = this.cache.get(contentType) || this.getParser(contentType)
+
+    if (parser === undefined) {
+      reply.send(new FST_ERR_CTP_INVALID_MEDIA_TYPE(contentType))
+    } else if (parser.asString === true || parser.asBuffer === true) {
+      rawBody(
+        request,
+        reply,
+        reply.context._parserOptions,
+        parser,
+        done
+      )
     } else {
-      request.body = body
-      handler(request, reply)
+      let result
+
+      if (parser.isDeprecatedSignature) {
+        result = parser.fn(request[kRequestPayloadStream], done)
+      } else {
+        result = parser.fn(request, request[kRequestPayloadStream], done)
+      }
+
+      if (result && typeof result.then === 'function') {
+        result.then(body => done(null, body), done)
+      }
+    }
+
+    function done (error, body) {
+      if (error) {
+        reply.send(error)
+      } else {
+        request.body = body
+        handler(request, reply)
+      }
     }
   }
 }