--- a/lib/hooks.js
+++ b/lib/hooks.js
@@ -30,32 +30,34 @@ const {
   kHooks
 } = require('./symbols')
 
-function Hooks () {
-  this.onRequest = []
-  this.preParsing = []
-  this.preValidation = []
-  this.preSerialization = []
-  this.preHandler = []
-  this.onResponse = []
-  this.onSend = []
-  this.onError = []
-  this.onRoute = []
-  this.onRegister = []
-  this.onReady = []
-  this.onTimeout = []
-}
-
-Hooks.prototype.validate = function (hook, fn) {
-  if (typeof hook !== 'string') throw new FST_ERR_HOOK_INVALID_TYPE()
-  if (typeof fn !== 'function') throw new FST_ERR_HOOK_INVALID_HANDLER()
-  if (supportedHooks.indexOf(hook) === -1) {
-    throw new Error(`${hook} hook not supported!`)
+class Hooks {
+  constructor () {
+    this.onRequest = []
+    this.preParsing = []
+    this.preValidation = []
+    this.preSerialization = []
+    this.preHandler = []
+    this.onResponse = []
+    this.onSend = []
+    this.onError = []
+    this.onRoute = []
+    this.onRegister = []
+    this.onReady = []
+    this.onTimeout = []
   }
-}
 
-Hooks.prototype.add = function (hook, fn) {
-  this.validate(hook, fn)
-  this[hook].push(fn)
+  validate (hook, fn) {
+    if (typeof hook !== 'string') throw new FST_ERR_HOOK_INVALID_TYPE()
+    if (typeof fn !== 'function') throw new FST_ERR_HOOK_INVALID_HANDLER()
+    if (supportedHooks.indexOf(hook) === -1) {
+      throw new Error(`${hook} hook not supported!`)
+    }
+  }
+
+  add (hook, fn) {
+    this.validate(hook, fn)
+    this[hook].push(fn)
+  }
 }
 
 function buildHooks (h) {