--- a/lib/reply.js
+++ b/lib/reply.js
@@ -50,20 +50,256 @@ const {
 } = require('./errors')
 const warning = require('./warnings')
 
-function Reply (res, request, log) {
-  this.raw = res
-  this[kReplySent] = false
-  this[kReplySerializer] = null
-  this[kReplyErrorHandlerCalled] = false
-  this[kReplyIsError] = false
-  this[kReplyIsRunningOnErrorHook] = false
-  this.request = request
-  this[kReplyHeaders] = {}
-  this[kReplyHasStatusCode] = false
-  this[kReplyStartTime] = undefined
-  this.log = log
+class Reply {
+  constructor (res, request, log) {
+    this.raw = res
+    this[kReplySent] = false
+    this[kReplySerializer] = null
+    this[kReplyErrorHandlerCalled] = false
+    this[kReplyIsError] = false
+    this[kReplyIsRunningOnErrorHook] = false
+    this.request = request
+    this[kReplyHeaders] = {}
+    this[kReplyHasStatusCode] = false
+    this[kReplyStartTime] = undefined
+    this.log = log
+  }
+
+  hijack () {
+    this[kReplySent] = true
+    return this
+  }
+
+  send (payload) {
+    if (this[kReplyIsRunningOnErrorHook] === true) {
+      throw new FST_ERR_SEND_INSIDE_ONERR()
+    }
+
+    if (this[kReplySent]) {
+      this.log.warn({ err: new FST_ERR_REP_ALREADY_SENT() }, 'Reply already sent')
+      return this
+    }
+
+    if (payload instanceof Error || this[kReplyIsError] === true) {
+      onErrorHook(this, payload, onSendHook)
+      return this
+    }
+
+    if (payload === undefined) {
+      onSendHook(this, payload)
+      return this
+    }
+
+    const contentType = this.getHeader('content-type')
+    const hasContentType = contentType !== undefined
+
+    if (payload !== null) {
+      if (Buffer.isBuffer(payload) || typeof payload.pipe === 'function') {
+        if (hasContentType === false) {
+          this[kReplyHeaders]['content-type'] = CONTENT_TYPE.OCTET
+        }
+        onSendHook(this, payload)
+        return this
+      }
+
+      if (hasContentType === false && typeof payload === 'string') {
+        this[kReplyHeaders]['content-type'] = CONTENT_TYPE.PLAIN
+        onSendHook(this, payload)
+        return this
+      }
+    }
+
+    if (this[kReplySerializer] !== null) {
+      if (typeof payload !== 'string') {
+        preserializeHook(this, payload)
+        return this
+      } else {
+        payload = this[kReplySerializer](payload)
+      }
+
+    // The indexOf below also matches custom json mimetypes such as 'application/hal+json' or 'application/ld+json'
+    } else if (hasContentType === false || contentType.indexOf('json') > -1) {
+      if (hasContentType === false) {
+        this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
+      } else {
+        // If hasContentType === true, we have a JSON mimetype
+        if (contentType.indexOf('charset') === -1) {
+          // If we have simply application/json instead of a custom json mimetype
+          if (contentType.indexOf('/json') > -1) {
+            this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
+          } else {
+            const currContentType = this[kReplyHeaders]['content-type']
+            // We extract the custom mimetype part (e.g. 'hal+' from 'application/hal+json')
+            const customJsonType = currContentType.substring(
+              currContentType.indexOf('/'),
+              currContentType.indexOf('json') + 4
+            )
+
+            // We ensure we set the header to the proper JSON content-type if necessary
+            // (e.g. 'application/hal+json' instead of 'application/json')
+            this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON.replace('/json', customJsonType)
+          }
+        }
+      }
+      if (typeof payload !== 'string') {
+        preserializeHook(this, payload)
+        return this
+      }
+    }
+
+    onSendHook(this, payload)
+
+    return this
+  }
+
+  getHeader (key) {
+    key = key.toLowerCase()
+    const res = this.raw
+    let value = this[kReplyHeaders][key]
+    if (value === undefined && res.hasHeader(key)) {
+      value = res.getHeader(key)
+    }
+    return value
+  }
+
+  getHeaders () {
+    return {
+      ...this.raw.getHeaders(),
+      ...this[kReplyHeaders]
+    }
+  }
+
+  hasHeader (key) {
+    key = key.toLowerCase()
+    if (this[kReplyHeaders][key] !== undefined) {
+      return true
+    }
+    return this.raw.hasHeader(key)
+  }
+
+  removeHeader (key) {
+    // Node.js does not like headers with keys set to undefined,
+    // so we have to delete the key.
+    delete this[kReplyHeaders][key.toLowerCase()]
+    return this
+  }
+
+  header (key, value) {
+    const _key = key.toLowerCase()
+
+    // default the value to ''
+    value = value === undefined ? '' : value
+
+    if (this[kReplyHeaders][_key] && _key === 'set-cookie') {
+      // https://tools.ietf.org/html/rfc7230#section-3.2.2
+      if (typeof this[kReplyHeaders][_key] === 'string') {
+        this[kReplyHeaders][_key] = [this[kReplyHeaders][_key]]
+      }
+      if (Array.isArray(value)) {
+        Array.prototype.push.apply(this[kReplyHeaders][_key], value)
+      } else {
+        this[kReplyHeaders][_key].push(value)
+      }
+    } else {
+      this[kReplyHeaders][_key] = value
+    }
+    return this
+  }
+
+  headers (headers) {
+    const keys = Object.keys(headers)
+    /* eslint-disable no-var */
+    for (var i = 0; i !== keys.length; ++i) {
+      const key = keys[i]
+      this.header(key, headers[key])
+    }
+    return this
+  }
+
+  code (code) {
+    const intValue = parseInt(code)
+    if (isNaN(intValue) || intValue < 100 || intValue > 600) {
+      throw new FST_ERR_BAD_STATUS_CODE(code || String(code))
+    }
+
+    this.raw.statusCode = intValue
+    this[kReplyHasStatusCode] = true
+    return this
+  }
+
+  serialize (payload) {
+    if (this[kReplySerializer] !== null) {
+      return this[kReplySerializer](payload)
+    } else {
+      if (this.context && this.context[kReplySerializerDefault]) {
+        return this.context[kReplySerializerDefault](payload, this.raw.statusCode)
+      } else {
+        return serialize(this.context, payload, this.raw.statusCode)
+      }
+    }
+  }
+
+  serializer (fn) {
+    this[kReplySerializer] = fn
+    return this
+  }
+
+  type (type) {
+    this[kReplyHeaders]['content-type'] = type
+    return this
+  }
+
+  redirect (code, url) {
+    if (typeof code === 'string') {
+      url = code
+      code = this[kReplyHasStatusCode] ? this.raw.statusCode : 302
+    }
+
+    this.header('location', url).code(code).send()
+  }
+
+  callNotFound () {
+    notFound(this)
+  }
+
+  getResponseTime () {
+    let responseTime = 0
+
+    if (this[kReplyStartTime] !== undefined) {
+      responseTime = now() - this[kReplyStartTime]
+    }
+
+    return responseTime
+  }
+
+  // Make reply a thenable, so it could be used with async/await.
+  // See
+  // - https://github.com/fastify/fastify/issues/1864 for the discussions
+  // - https://promisesaplus.com/ for the definition of thenable
+  // - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/then for the signature
+  then (fulfilled, rejected) {
+    if (this.sent) {
+      fulfilled()
+      return
+    }
+
+    eos(this.raw, (err) => {
+      // We must not treat ERR_STREAM_PREMATURE_CLOSE as
+      // an error because it is created by eos, not by the stream.
+      if (err && err.code !== 'ERR_STREAM_PREMATURE_CLOSE') {
+        if (rejected) {
+          rejected(err)
+        } else {
+          this.log && this.log.warn('unhandled rejection on reply.then')
+        }
+      } else {
+        fulfilled()
+      }
+    })
+  }
 }
 
+Reply.prototype.status = Reply.prototype.code
 Object.defineProperties(Reply.prototype, {
   context: {
     get () {
@@ -107,241 +343,6 @@ Object.defineProperties(Reply.prototype, {
     writable: true
   }
 })
-
-Reply.prototype.hijack = function () {
-  this[kReplySent] = true
-  return this
-}
-
-Reply.prototype.send = function (payload) {
-  if (this[kReplyIsRunningOnErrorHook] === true) {
-    throw new FST_ERR_SEND_INSIDE_ONERR()
-  }
-
-  if (this[kReplySent]) {
-    this.log.warn({ err: new FST_ERR_REP_ALREADY_SENT() }, 'Reply already sent')
-    return this
-  }
-
-  if (payload instanceof Error || this[kReplyIsError] === true) {
-    onErrorHook(this, payload, onSendHook)
-    return this
-  }
-
-  if (payload === undefined) {
-    onSendHook(this, payload)
-    return this
-  }
-
-  const contentType = this.getHeader('content-type')
-  const hasContentType = contentType !== undefined
-
-  if (payload !== null) {
-    if (Buffer.isBuffer(payload) || typeof payload.pipe === 'function') {
-      if (hasContentType === false) {
-        this[kReplyHeaders]['content-type'] = CONTENT_TYPE.OCTET
-      }
-      onSendHook(this, payload)
-      return this
-    }
-
-    if (hasContentType === false && typeof payload === 'string') {
-      this[kReplyHeaders]['content-type'] = CONTENT_TYPE.PLAIN
-      onSendHook(this, payload)
-      return this
-    }
-  }
-
-  if (this[kReplySerializer] !== null) {
-    if (typeof payload !== 'string') {
-      preserializeHook(this, payload)
-      return this
-    } else {
-      payload = this[kReplySerializer](payload)
-    }
-
-  // The indexOf below also matches custom json mimetypes such as 'application/hal+json' or 'application/ld+json'
-  } else if (hasContentType === false || contentType.indexOf('json') > -1) {
-    if (hasContentType === false) {
-      this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
-    } else {
-      // If hasContentType === true, we have a JSON mimetype
-      if (contentType.indexOf('charset') === -1) {
-        // If we have simply application/json instead of a custom json mimetype
-        if (contentType.indexOf('/json') > -1) {
-          this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
-        } else {
-          const currContentType = this[kReplyHeaders]['content-type']
-          // We extract the custom mimetype part (e.g. 'hal+' from 'application/hal+json')
-          const customJsonType = currContentType.substring(
-            currContentType.indexOf('/'),
-            currContentType.indexOf('json') + 4
-          )
-
-          // We ensure we set the header to the proper JSON content-type if necessary
-          // (e.g. 'application/hal+json' instead of 'application/json')
-          this[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON.replace('/json', customJsonType)
-        }
-      }
-    }
-    if (typeof payload !== 'string') {
-      preserializeHook(this, payload)
-      return this
-    }
-  }
-
-  onSendHook(this, payload)
-
-  return this
-}
-
-Reply.prototype.getHeader = function (key) {
-  key = key.toLowerCase()
-  const res = this.raw
-  let value = this[kReplyHeaders][key]
-  if (value === undefined && res.hasHeader(key)) {
-    value = res.getHeader(key)
-  }
-  return value
-}
-
-Reply.prototype.getHeaders = function () {
-  return {
-    ...this.raw.getHeaders(),
-    ...this[kReplyHeaders]
-  }
-}
-
-Reply.prototype.hasHeader = function (key) {
-  key = key.toLowerCase()
-  if (this[kReplyHeaders][key] !== undefined) {
-    return true
-  }
-  return this.raw.hasHeader(key)
-}
-
-Reply.prototype.removeHeader = function (key) {
-  // Node.js does not like headers with keys set to undefined,
-  // so we have to delete the key.
-  delete this[kReplyHeaders][key.toLowerCase()]
-  return this
-}
-
-Reply.prototype.header = function (key, value) {
-  const _key = key.toLowerCase()
-
-  // default the value to ''
-  value = value === undefined ? '' : value
-
-  if (this[kReplyHeaders][_key] && _key === 'set-cookie') {
-    // https://tools.ietf.org/html/rfc7230#section-3.2.2
-    if (typeof this[kReplyHeaders][_key] === 'string') {
-      this[kReplyHeaders][_key] = [this[kReplyHeaders][_key]]
-    }
-    if (Array.isArray(value)) {
-      Array.prototype.push.apply(this[kReplyHeaders][_key], value)
-    } else {
-      this[kReplyHeaders][_key].push(value)
-    }
-  } else {
-    this[kReplyHeaders][_key] = value
-  }
-  return this
-}
-
-Reply.prototype.headers = function (headers) {
-  const keys = Object.keys(headers)
-  /* eslint-disable no-var */
-  for (var i = 0; i !== keys.length; ++i) {
-    const key = keys[i]
-    this.header(key, headers[key])
-  }
-  return this
-}
-
-Reply.prototype.code = function (code) {
-  const intValue = parseInt(code)
-  if (isNaN(intValue) || intValue < 100 || intValue > 600) {
-    throw new FST_ERR_BAD_STATUS_CODE(code || String(code))
-  }
-
-  this.raw.statusCode = intValue
-  this[kReplyHasStatusCode] = true
-  return this
-}
-
-Reply.prototype.status = Reply.prototype.code
-
-Reply.prototype.serialize = function (payload) {
-  if (this[kReplySerializer] !== null) {
-    return this[kReplySerializer](payload)
-  } else {
-    if (this.context && this.context[kReplySerializerDefault]) {
-      return this.context[kReplySerializerDefault](payload, this.raw.statusCode)
-    } else {
-      return serialize(this.context, payload, this.raw.statusCode)
-    }
-  }
-}
-
-Reply.prototype.serializer = function (fn) {
-  this[kReplySerializer] = fn
-  return this
-}
-
-Reply.prototype.type = function (type) {
-  this[kReplyHeaders]['content-type'] = type
-  return this
-}
-
-Reply.prototype.redirect = function (code, url) {
-  if (typeof code === 'string') {
-    url = code
-    code = this[kReplyHasStatusCode] ? this.raw.statusCode : 302
-  }
-
-  this.header('location', url).code(code).send()
-}
-
-Reply.prototype.callNotFound = function () {
-  notFound(this)
-}
-
-Reply.prototype.getResponseTime = function () {
-  let responseTime = 0
-
-  if (this[kReplyStartTime] !== undefined) {
-    responseTime = now() - this[kReplyStartTime]
-  }
-
-  return responseTime
-}
-
-// Make reply a thenable, so it could be used with async/await.
-// See
-// - https://github.com/fastify/fastify/issues/1864 for the discussions
-// - https://promisesaplus.com/ for the definition of thenable
-// - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/then for the signature
-Reply.prototype.then = function (fulfilled, rejected) {
-  if (this.sent) {
-    fulfilled()
-    return
-  }
-
-  eos(this.raw, (err) => {
-    // We must not treat ERR_STREAM_PREMATURE_CLOSE as
-    // an error because it is created by eos, not by the stream.
-    if (err && err.code !== 'ERR_STREAM_PREMATURE_CLOSE') {
-      if (rejected) {
-        rejected(err)
-      } else {
-        this.log && this.log.warn('unhandled rejection on reply.then')
-      }
-    } else {
-      fulfilled()
-    }
-  })
-}
 
 function preserializeHook (reply, payload) {
   if (reply.context.preSerialization !== null) {