--- a/lib/request.js
+++ b/lib/request.js
@@ -4,14 +4,16 @@ const proxyAddr = require('proxy-addr')
 const semver = require('semver')
 const warning = require('./warnings')
 
-function Request (id, params, req, query, log, context) {
-  this.id = id
-  this.context = context
-  this.params = params
-  this.raw = req
-  this.query = query
-  this.log = log
-  this.body = null
+class Request {
+  constructor (id, params, req, query, log, context) {
+    this.id = id
+    this.context = context
+    this.params = params
+    this.raw = req
+    this.query = query
+    this.log = log
+    this.body = null
+  }
 }
 
 function getTrustProxyFn (tp) {