--- a/lib/schemas.js
+++ b/lib/schemas.js
@@ -12,35 +12,37 @@ const {
 
 const SCHEMAS_SOURCE = ['params', 'body', 'querystring', 'query', 'headers']
 
-function Schemas (initStore) {
-  this.store = initStore || {}
-}
-
-Schemas.prototype.add = function (inputSchema) {
-  const schema = fastClone((inputSchema.isFluentSchema || inputSchema.isFluentJSONSchema || inputSchema[kFluentSchema])
-    ? inputSchema.valueOf()
-    : inputSchema
-  )
-
-  // devs can add schemas without $id, but with $def instead
-  const id = schema.$id
-  if (!id) {
-    throw new FST_ERR_SCH_MISSING_ID()
+class Schemas {
+  constructor (initStore) {
+    this.store = initStore || {}
   }
 
-  if (this.store[id]) {
-    throw new FST_ERR_SCH_ALREADY_PRESENT(id)
+  add (inputSchema) {
+    const schema = fastClone((inputSchema.isFluentSchema || inputSchema.isFluentJSONSchema || inputSchema[kFluentSchema])
+      ? inputSchema.valueOf()
+      : inputSchema
+    )
+
+    // devs can add schemas without $id, but with $def instead
+    const id = schema.$id
+    if (!id) {
+      throw new FST_ERR_SCH_MISSING_ID()
+    }
+
+    if (this.store[id]) {
+      throw new FST_ERR_SCH_ALREADY_PRESENT(id)
+    }
+
+    this.store[id] = schema
   }
 
-  this.store[id] = schema
-}
+  getSchemas () {
+    return Object.assign({}, this.store)
+  }
 
-Schemas.prototype.getSchemas = function () {
-  return Object.assign({}, this.store)
-}
-
-Schemas.prototype.getSchema = function (schemaId) {
-  return this.store[schemaId]
+  getSchema (schemaId) {
+    return this.store[schemaId]
+  }
 }
 
 function normalizeSchema (routeSchemas) {