<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

This PR only introduces a syntax change; it refactors constructor functions and modifying prototypes to using ES6 classes. I ran benchmarks a few times there isn't any noticable difference since benchmarks are very inconsistent.

These were ran using nodejs v10.17.0.
"main"

```
[1] ┌─────────┬──────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼──────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 9 ms │ 20 ms │ 41 ms │ 61 ms │ 19.67 ms │ 10.3 ms │ 182 ms │
[1] └─────────┴──────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 35263   │ 35263   │ 55583   │ 56063  │ 49622.4 │ 8279.23 │ 35239   │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 5.78 MB │ 5.78 MB │ 9.12 MB │ 9.2 MB │ 8.14 MB │ 1.36 MB │ 5.78 MB │
[1] └───────────┴─────────┴─────────┴─────────┴────────┴─────────┴─────────┴─────────┘
```

```
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 10 ms │ 21 ms │ 43 ms │ 63 ms │ 20.41 ms │ 10.25 ms │ 176 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 34623   │ 34623   │ 52927   │ 53983   │ 47856   │ 7489.29 │ 34616   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 5.68 MB │ 5.68 MB │ 8.68 MB │ 8.86 MB │ 7.85 MB │ 1.23 MB │ 5.68 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```

"pr/refactor-to-es6-classes"

```
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 10 ms │ 22 ms │ 43 ms │ 59 ms │ 21.05 ms │ 10.93 ms │ 151 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 34047   │ 34047   │ 50143   │ 52511   │ 46460.8 │ 7062.01 │ 34021   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 5.58 MB │ 5.58 MB │ 8.22 MB │ 8.61 MB │ 7.62 MB │ 1.16 MB │ 5.58 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```

```
[1] ┌─────────┬──────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼──────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 9 ms │ 20 ms │ 37 ms │ 60 ms │ 19.04 ms │ 9.52 ms │ 174 ms │
[1] └─────────┴──────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 36895   │ 36895   │ 54527   │ 58079   │ 51190.4 │ 7966.28 │ 36895   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 6.05 MB │ 6.05 MB │ 8.95 MB │ 9.53 MB │ 8.4 MB  │ 1.31 MB │ 6.05 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘
```