@sergiz 2021-06-29
It's actually a tiny change. Github makes it look gigantic, but if you hide whitespace changes there were only a few lines actually modified.

It makes the code utilize the modern Javascript and it's easier to understand at least to those who started with JS after ES6 was out.