@zekth 2021-06-29
> It complicates the code without any tangible benefit.

It doesn't complicates, it uses the modern ES syntax. As stated from [MDN ](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes): 

> Classes are in fact "special functions", and just as you can define function expressions and function declarations, the class syntax has two components: class expressions and class declarations.

I have no strong opinion on this, i'd say it's easier to read and onboard new developers. One benefit would be the [Private class features](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes/Private_class_fields).