@jsumners 2021-06-29
> It doesn't complicates, it uses the modern ES syntax.

The extent of my usage of the `class` keyword in JavaScript is literally summed up by:

```js
'use strict'

const MySet = class extends Set {}

MySet.prototype.find = function find(value) {
  for (const item of this) {
    if (item === value) return item
  }
}

const mySet = new MySet(['foo', 'bar', 'baz'])
console.log(mySet.find('bar'))
```

And that is because the spec requires such nonsense -- https://tc39.es/ecma262/#sec-set-constructor

Otherwise, I am not at all clear what other trickery the `class` keyword introduces that would screw with the typical prototype methodology and find it more difficult to read/work with.