@MSE99 2021-06-29
Methods in the class bodies suffer from an extra level of indentation, this is the only negative i see, classes have a lot of differences to regular functions but i doubt that any of them is going to break fastify or one of the plugins.