@L2jLiga 2021-06-30
I'm +1 on that. According to my experience IDEs better handle classes than prototypes