@RafaelGSS 2021-06-30
> Most of the "classes" touched by this are used in complex prototype chains and acts together with our encapsulation behaviors. Using a class sends the wrong signal.

Our tests would break if there are any key errors on it, no? I mean, I would like to see a canary release with that refactor, I see a huge benefit for newcomers.