@Eomm 2021-07-01
Maybe could we think to use this syntax for those classes that don't use the prototype inheritance?

I mean:
- schemas
- hooks
- content type parser