@jsumners 2021-07-01
> Maybe could we think to use this syntax for those classes that don't use the prototype inheritance?

I don't see any reason to mix and match syntax. It would introduce nebulous rules that would be a mystery to contributors and difficult to remember for maintainers.