@Eomm 2021-07-01
I feel guilty right now 😄 

We have a class

https://github.com/fastify/fastify/blob/e4b2a3aad3a4e56ab3c9192aa6bb989316ef7780/lib/schema-controller.js#L29

Have I [misunderstood](https://github.com/fastify/fastify/pull/2755#pullrequestreview-558878398) this review?