@mcollina 2021-08-03
This seems a bit stuck without significant progress. I'm going to close it for now.

Feel free to reopen if you are still interested in working on this.